+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - marked class Shader as deprecated, use class OpenGLShaderProgram instead!
   - DefaultCamera: added support to detect, if camera was changed (for example position, rotation and so on), so renderer can be optimized and dont have to generate some matrizes every render cycle
   - FloatVertexBufferObject & IntegerVertexBufferObject now supports GL_DYNAMIC_DRAW
+  - OpenGLShaderProgram now also supports Vector4f uniform
 
 Special thanks to noctarius for some tips!